+++ release-notes.md
@@ -1,3 +1,8 @@
+# v------
+*   Adding `q` to quit the application immediately.  You no longer have to
+    back all of the way out.
+***
+
 # v1.1.2
 *   FIX: setup.py not compatible with Python 3.
 ***